Enhance error handling and add versioncheck to neovim

+++ packages/neovim.sh
@@ -1,9 +1,21 @@
 #!/usr/bin/env bash
 
-BASEDIR="$(dirname "$(readlink -f "$0")")"
-TMPDIR="$(mktemp -d)"
-
 set -eu -o pipefail
+
+BASEDIR="$(dirname "$(readlink -f "$0")")"
+RELEASE_URL="https://api.github.com/repos/neovim/neovim/releases/latest"
+
+versioncheck () {
+	set +eu
+	localversion="$(nvim --version 2> /dev/null | grep -oP '(?<=NVIM ).*(?=$)')"
+	set -eu
+	remoteversion="$(curl -s "$RELEASE_URL" | jq -r ".tag_name")"
+	if [ "$localversion" = "$remoteversion" ] ; then
+		echo "nvim is up to date"
+		exit 0
+	fi
+}
+versioncheck
 
 #################################################################################
 # setup exit handler
@@ -11,7 +23,9 @@ set -eu -o pipefail
 onexit() {
 	cd "$BASEDIR"
 	echo "Script is terminating -- cleaning up"
-	sudo rm -rf "$TMPDIR"
+	if [ -d "$TMPDIR" ] ; then
+		sudo rm -rf "$TMPDIR"
+	fi
 	exit
 }
 
@@ -21,7 +35,8 @@ trap '' INT TERM # Ignore SigINT and SigTERM
 sudo apt update
 sudo apt install -y jq curl libtool libtool-bin autoconf automake cmake g++ pkg-config unzip python3-pip
 
-URL="$(curl -s "https://api.github.com/repos/neovim/neovim/releases/latest" | jq -r ".zipball_url")"
+TMPDIR="$(mktemp -d)"
+URL="$(curl -s "$RELEASE_URL" | jq -r ".zipball_url")"
 
 cd "$TMPDIR"
 curl -L -o neovim.zip "$URL"
@@ -32,17 +47,7 @@ make -j"$(nproc --all)" CMAKE_BUILD_TYPE=Release
 sudo make install
 pip3 install --upgrade neovim
 
-set +e
-grep 'alias vim="nvim"' "$HOME/.bashrc.local" &> /dev/null
-set -e
-
-if [ "$?" != "0" ] ; then
-	echo 'alias vim="nvim -u ~/.vimrc"' >> "$HOME/.bashrc.local"
-	echo 'alias vi="vim"' >> "$HOME/.bashrc.local"
-fi
-
 if [ ! -e "$HOME/.config/nvim/" ] ; then
 	cp -r "$HOME/.vim" "$HOME/.config/nvim"
 	cp "$HOME/.vimrc" "$HOME/.config/nvim/init.vim"
 fi
-
